Changed from MySQL database to built-in Replit Database. Added mute.py, updated prefix.py, and updated filter.py for more robust functionality of the discord bot.

+++ cogs/filter.py
@@ -10,7 +10,6 @@ class Filter(commands.Cog):
   def __init__(self, bot):
     self.bot = bot
     self.server_id = None
-    self.msg_id = None
   
   # Displays when cog has been loaded in
   @commands.Cog.listener()
@@ -21,27 +20,51 @@ class Filter(commands.Cog):
   @commands.Cog.listener()
   async def on_message(self, message):
 
-    # Reads in user message and loads in profanity csv
-    msg = message.content
-    df = pd.read_csv('https://query.data.world/s/o5363h3dmwli3yuolcxwg52llwxpmh')
+    # Gets server ID
+    self.server_id = message.guild.id
+    
+    # Checks to see if filter is 'ON'
+    if db[str(self.server_id)][0]["profanity toggle"] == "ON":
+      # Reads in user message and loads in profanity csv
+      msg = message.content
+      df = pd.read_csv('https://query.data.world/s/o5363h3dmwli3yuolcxwg52llwxpmh')
 
-    # Checks to see if message sender was the bot itself
-    if message.author == self.bot.user:
-      return
+      # Checks to see if message sender was the bot itself
+      if message.author == self.bot.user:
+        return
 
-    # Deletes and responds to profanity by the offender
-    msg = msg.split()
+      # Deletes and responds to profanity by the offender
+      msg = msg.split()
 
-    if msg in df.values:
-      await message.delete()
-      embed = discord.Embed(title="Curse Filter", description="Your messsage contained profanity and was deleted. Please review the rules of the server before messaging!", color=0xFF5733)
+      if msg in df.values:
+        await message.delete()
+        embed = discord.Embed(title="Curse Filter", description="Your messsage contained profanity and was deleted. Please review the rules of the server before messaging!", color=0xFF5733)
 
-      msg = await message.channel.send(embed=embed)
+        msg = await message.channel.send(embed=embed)
 
-      # Deletes message after 3 seconds
-      sleep(3)
-      await msg.delete()
+        # Deletes message after 3 seconds
+        sleep(3)
+        await msg.delete()
 
     
+  @commands.command(name='filter',description="Toggles the usage of the profanity filter")
+  @commands.has_permissions(administrator=True)
+  async def filtertoggle(self, ctx):
+    
+    # Gets Server ID
+    self.server_id = ctx.guild.id
+
+    # Checks servers current filter preference
+    pref = db[str(self.server_id)][0]["profanity toggle"]
+
+    # toggles preference
+    if pref == "ON":
+      db[str(self.server_id)][0]["profanity toggle"] = "OFF"
+    
+    else:
+      db[str(self.server_id)][0]["profanity toggle"] = "ON"
+
+    await ctx.send("Profanity filter is now set to '"+ db[str(self.server_id)][0]["profanity toggle"] + "'")
+
 def setup(bot):
   bot.add_cog(Filter(bot))
